Reorganize utilities and start events
