+++ plugins/treeview/web_client/children/collections.js
@@ -1,6 +1,6 @@
 import _ from 'underscore';
 
-import request from '../request';
+import request from '../utils/request';
 import { collection } from '../types';
 
 export default function () {
