+++ plugins/treeview/web_client/children/folder.js
@@ -1,6 +1,6 @@
 import _ from 'underscore';
 
-import request from '../request';
+import request from '../utils/request';
 import { folder, item } from '../types';
 
 export default function (parent) {
