+++ plugins/treeview/web_client/children/users.js
@@ -1,14 +1,13 @@
 import _ from 'underscore';
 
-import request from '../request';
-import getCurrentUser from '../auth';
+import { request, auth } from '../utils';
 import { user } from '../types';
 
 export default function () {
     return request({
         path: 'user'
     }).then((users) => {
-        const current = getCurrentUser() || {};
+        const current = auth() || {};
         return _.map(
             _.reject(users, {_id: current._id}),
             user
