+++ plugins/treeview/web_client/index.js
@@ -1,15 +1,11 @@
-import auth from './auth';
-import request from './request';
-import attach from './attach';
+import * as utils from './utils';
 import * as types from './types';
 import * as children from './children';
 import * as views from './views';
 
 export {
     views,
-    attach,
-    auth,
-    request,
+    utils,
     types,
     children
 };
