+++ plugins/treeview/web_client/views/TreeView.js
@@ -3,6 +3,10 @@ import View from 'girder/views/View';
 import attach from '../attach';
 
 const TreeView = View.extend({
+    events: {
+        'changed.jstree': '_onChanged'
+    },
+
     initialize(settings) {
     },
 
@@ -16,8 +20,20 @@ const TreeView = View.extend({
         return View.prototype.destroy.apply(this, arguments);
     },
 
+    instance() {
+        return this.$el.jstree(true);
+    },
+
     _destroy() {
         this.$el.jstree('destroy');
+    },
+
+    _onChanged(e, data) {
+        this.trigger(
+            'g:treeview:select', {
+
+            }
+        );
     }
 });
